+++ test/test_cli.py
@@ -93,7 +93,9 @@ class CliTestCase(unittest.TestCase):
                 response,
                 command,
                 default_category=configuration.get_option(
-                    "FRONTEND", "default_category"))
+                    "FRONTEND", "default_category"),
+                recurrent_only=params.get("recurrent_only", False),
+            )
 
         return response
 
@@ -247,6 +249,18 @@ default_category = no-category"""
         response = self.cli_run("list -f value=10").lower()
         self.assertIn("money", response)
         self.assertIn("10.00", response)
+
+    def test_list_recurrent_only(self):
+        self.cli_run("add interest 20 -s 2020-01-01 -f yearly -c banking")
+        self.cli_run("add rent -300 -s 2020-06-15 -e 2021-12-31 -f monthly")
+        response = self.cli_run("list --recurrent-only")
+        # yapf: disable
+        self.assertEqual(
+            response,
+        "ID | NAME     | VALUE  | CATEGORY    | START      | END        | FREQUENCY\n" +  # noqa
+        " 1 | Interest |   20.0 | Banking     | 2020-01-01 | -          | Yearly   \n" +  # noqa
+        " 2 | Rent     | -300.0 | Unspecified | 2020-06-15 | 2021-12-31 | Monthly  ") # noqa
+        # yapf: enable
 
     @mock.patch("financeager.server.Server.run")
     def test_communication_error(self, mocked_run):